:recycle: Move createSafeObjectInputStream to java 8 compatible class
