+++ src/java11/main/io/github/pixee/security/ObjectInputFilters.java
@@ -1,14 +1,12 @@
 package io.github.pixee.security;
 
-import java.io.IOException;
-import java.io.InputStream;
 import java.io.ObjectInputFilter;
 import java.io.ObjectInputStream;
 import java.util.Objects;
-import org.apache.commons.io.serialization.ValidatingObjectInputStream;
 
 /**
- * This type exposes helper methods that will help defend against Java deserialization attacks.
+ * This type exposes helper methods that will help defend against Java deserialization attacks
+ * leveraging the Java 9+ API {@link ObjectInputFilter}.
  *
  * <p>For more information on deserialization checkout the <a
  * href="https://cheatsheetseries.owasp.org/cheatsheets/Deserialization_Cheat_Sheet.html">OWASP
@@ -92,30 +90,6 @@ public final class ObjectInputFilters {
       }
       return originalFilter.checkInput(filterInfo);
     }
-  }
-
-  /**
-   * This method returns a wrapped {@link ObjectInputStream} that protects against deserialization
-   * code execution attacks. This method can be used in Java 8 and previous.
-   *
-   * @param ois the stream to wrap and harden
-   * @return an {@link ObjectInputStream} which is safe against all publicly known gadgets
-   * @throws IOException if the underlying creation of {@link ObjectInputStream} fails
-   */
-  public static ObjectInputStream createSafeObjectInputStream(final InputStream ois)
-      throws IOException {
-    try {
-      final ValidatingObjectInputStream is = new ValidatingObjectInputStream(ois);
-      for (String gadget : UnwantedTypes.dangerousClassNameTokens()) {
-        is.reject("*" + gadget + "*");
-      }
-      return is;
-    } catch (IOException e) {
-      // ignored
-    }
-
-    // if for some reason we can't replace it, we'll pass it back as it was given
-    return new ObjectInputStream(ois);
   }
 
   private static final ObjectInputFilter basicGadgetDenylistFilter =
